agents/abrir_conta/requirements.txt
a2a-sdk[http-server]
httpx
pydantic
uvicorn
starlette
langchain
langchain_openai
langchain-community
python-dotenv